+++ app/controllers/forms/form_get_interface_ae0_summary.py
@@ -1,0 +1,11 @@
+from flask_wtf import FlaskForm
+from wtforms import SelectField, SubmitField
+
+
+class SummaryForm(FlaskForm):
+    hostname = SelectField(
+        'Hostname', choices=[],
+    )
+    submit = SubmitField(
+        'Enviar'
+    )
