+++ app/templates/vendors/junos/get_interface_ae0_summary.html
@@ -1,43 +1,18 @@
 {% extends "base.html" %}
 
 {% block content %}
-<h1 class="display-4">{% block title %}Get Summary{% endblock %}</h1>
-
-{% block comment %}
-{{ super() }}
-{% endblock %}
+<h1>{% block title %}Get Summary{% endblock %}</h1>
 
 <hr>
 
 <form method="POST" action="{{ url_for('int_summary_bp.interface_summary') }}" autocomplete="on">
     {{ form.hidden_tag() }}
 
-    <div class="form-row">
-        <!-- Routers -->
-        <div class="form-group col">
-            <select name="hostname" id="hostname" class="form-control">
-                {% for device in devices %}
-                <option value="{{ device.ip_address }}">{{ device.hostname }}</option>
-                {% endfor %}
-            </select>
-        </div>
-    </div>
-
-    <div class="form-row">
-        <!-- Usuários -->
-        <div class="form-group col mb-3">
-            <input type="text" class="form-control" value="{{ current_user.username }}" readonly>
-        </div>
-
-        <!-- Passwords -->
-        <div class="input-group col mb-3">
-            <input type="password" class="form-control" id="password-input" value="{{ current_user.password[:12] }}"
-                readonly>
-            <div class="input-group-append">
-                <i class="fas fa-eye input-group-text" id="toggle-password"></i>
-            </div>
-        </div>
-    </div>
+    <select class="form-select">
+        <option selected>Open this select menu</option>
+        <option value="1">One</option>
+    </select>
+    <div class="form-text mb-3">We'll never share your email with anyone else.</div>
 
     <button type="submit" class="btn btn-primary">Enviar</button>
 </form>
@@ -45,7 +20,6 @@
 <br>
 
 <pre class="text-center">show interfaces terse lo0 | match lo</pre>
-
 <pre class="max-height" style=" border: solid 1px rgba(0, 0, 0, .1); 
                                     padding: 1rem; 
                                     margin-top: 1rem;
